refactor(codegen): extract form method dispatchers into form/mod.rs (Round 7b)
8 fns lifted as inherent impls on Cx into form/mod.rs (~2,106 lines):
- try_lower_form_vec_method + try_lower_form_vec_fallible_method
- try_lower_form_hashmap_method + try_lower_form_hashmap_fallible_method
- try_lower_form_ring_buffer_method + try_lower_form_ring_buffer_fallible_method
- lower_form_hashmap_index_method (the [] / get / set / has / remove
  shape)
- lower_form_hashmap_bump (the increment-or-init shape for
  cells like { key + one Int counter })

These are the @form(...) synthesized-method call dispatchers — the
piece the original plan called "Round 2" before scope absorption.
Slot allocation + per-form init shipped in Round 4d
(locus/instantiation.rs).

Surface bumps to pub(crate):
- Cx methods: emit_cross_arena_store_deep_copy, emit_key_error_alloc,
  emit_empty_error_alloc, emit_or_get_sort_trampoline,
  try_lower_form_vec_method, try_lower_form_hashmap_method,
  try_lower_form_ring_buffer_method, lower_form_hashmap_index_method,
  lower_form_hashmap_bump

codegen.rs: 22967 -> 20889 (-2078 lines).
